Clean up context menu items into hook
